+++ RazorSvelte/App/401.svelte
@@ -12,3 +12,10 @@
         </p>
     </div>
 </Layout>
+
+<style lang="scss">
+    :global(main) {
+        margin: auto;
+    }
+</style>
+