+++ RazorSvelte/App/404.svelte
@@ -11,3 +11,10 @@
         </p>
     </div>
 </Layout>
+
+<style lang="scss">
+    :global(main) {
+        margin: auto;
+    }
+</style>
+