+++ RazorSvelte/App/authorized-page.svelte
@@ -25,3 +25,9 @@
     </div>
 </Layout>
 
+<style lang="scss">
+    :global(main) {
+        margin: auto;
+    }
+</style>
+