+++ RazorSvelte/App/login.svelte
@@ -7,20 +7,30 @@
     <div class="container text-center m-5">
         <ul class="list-group">
             <li class="list-group-item">
-                <a class="btn btn-primary btn-lg" href="{urls.signInGoogleUrl}">Login With Google</a>
+                <a class="btn btn-primary btn-lg" href="{urls.signInGoogleUrl}">
+                    <i class="bi bi-google"></i>
+                    Login With Google
+                </a>
             </li>
             <li class="list-group-item">
-                <a class="btn btn-primary btn-lg" href="{urls.signInLinkedInUrl}">Login With LinkedIn</a>
+                <a class="btn btn-primary btn-lg" href="{urls.signInLinkedInUrl}">
+                    <i class="bi bi-linkedin"></i>
+                    Login With LinkedIn
+                </a>
             </li>
             <li class="list-group-item">
-                <a class="btn btn-primary btn-lg" href="{urls.signInGitHubUrl}">Login With GitHub</a>
+                <a class="btn btn-primary btn-lg" href="{urls.signInGitHubUrl}">
+                    <i class="bi bi-github"></i>
+                    Login With GitHub
+                </a>
             </li>
         </ul>
     </div>
 </Layout>
 
 <style lang="scss">
-    li {
-        border: 0;
+    :global(main) {
+        margin: auto;
     }
 </style>
+