+++ RazorSvelte/App/privacy.svelte
@@ -12,4 +12,10 @@
     </div>
 </Layout>
 
+<style lang="scss">
+:global(main) {
+    margin: auto;
+}
+</style>
+