+++ RazorSvelte/App/spa-example.svelte
@@ -5,49 +5,51 @@
     import About from "./spa-views/about.svelte";
     import Parametrized from "./spa-views/parametrized.svelte";
     import Unknown from "./spa-views/unknown.svelte";
+
+    let hash = document.location.hash;
+    window.onhashchange = () => hash = document.location.hash;
 </script>
 
 <Layout>
-    <div class="container">
-        <h1>Spa Example</h1>
+    <div class="container mt-3">
+
+        <h3>Spa Example</h3>
         <p>
             This is an example of the Single-Page App (SPA) by using 
-            <a class="btn btn-link" href="https://www.npmjs.com/package/svelte-spa-router"><code>svelte-spa-router</code></a> component.
-            <br />
-            See <a class="" href="https://svelte.dev/repl/6e7546119cff41fabe35df3ef999ff7e?version=3.29.4">REPL</a> and/or 
-            <a class="" href="https://github.com/ItalyPaleAle/svelte-spa-router">GitHub</a> for this component.
+            <a class="btn-link" href="https://www.npmjs.com/package/svelte-spa-router">svelte-spa-router</a> component.
+        </p>
+        <p>
+            See <a class="btn-link" href="https://svelte.dev/repl/6e7546119cff41fabe35df3ef999ff7e?version=3.29.4">REPL</a> and/or 
+            <a class="btn-link" href="https://github.com/ItalyPaleAle/svelte-spa-router">GitHub</a> for this component.
         </p>
 
-        <div class="container p-0">
+        <ul class="nav nav-tabs">
+            <li class="nav-item">
+                <!-- svelte-ignore a11y-invalid-attribute -->
+                <a class="nav-link" class:active={hash == ""} href="#">Home</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" class:active={hash == "#/about"} href="#/about">About</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" class:active={hash == "#/parametrized/AAA"} href="#/parametrized/AAA">Parametrized with "AAA"</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" class:active={hash == "#/parametrized/BBB"} href="#/parametrized/BBB">Parametrized with "BBB"</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#/test">Unknown route</a>
+            </li>
+        </ul>
 
-            <ul class="nav justify-content-center border">
-                <li class="nav-item">
-                    <!-- svelte-ignore a11y-invalid-attribute -->
-                    <a class="nav-link active" aria-current="page" href="#">Home</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="#/about">About</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="#/parametrized/AAA">Parametrized with "AAA"</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="#/parametrized/BBB">Parametrized with "BBB"</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="#/test">Unknown route</a>
-                </li>
-            </ul>
-
-            <div class="p-0 mt-5">
-                <Router routes={{
-                    "/": Home,
-                    "/about": About,
-                    "/parametrized/:param": Parametrized,
-                    "*": Unknown,
-                    //'/blog/:blogName': Blog
-                }} />
-            </div>
+        <div class="p-0 mt-3">
+            <Router routes={{
+                "/": Home,
+                "/about": About,
+                "/parametrized/:param": Parametrized,
+                "*": Unknown,
+                //'/blog/:blogName': Blog
+            }} />
         </div>
     </div>
 </Layout>