+++ RazorSvelte/Pages/AuthorizedPage.entry.ts
@@ -1,4 +1,3 @@
 /// <reference types="svelte" />
 import App from "../App/authorized-page.svelte";
-import "bootstrap/js/src/collapse";
 export default new App({target: document.body});