+++ RazorSvelte/Pages/Unauthorized.rollup.config.js
@@ -1,2 +1,2 @@
 import config from "../Scripts/rollup.config";
-export default config({ input: "./Pages/Unauthorized.entry.ts", jsOutput: "./wwwroot/build/401.js"});
+export default config({ input: "./Pages/Unauthorized.entry.ts", jsOutput: "./wwwroot/build/401.js", cssOutput: "401.css"});